chore: move sidetree operations to op
